NF-666 - improve code coverage - min set to 80% - deleted redundant code
