+++ app/uk/gov/hmrc/nationaldutyrepaymentcenter/models/requests/AmendClaimRequest.scala
@@ -17,7 +17,7 @@
 package uk.gov.hmrc.nationaldutyrepaymentcenter.models.requests
 
 import play.api.libs.json.{Format, Json}
-import uk.gov.hmrc.nationaldutyrepaymentcenter.models.{AmendContent, UploadedFile, Validator}
+import uk.gov.hmrc.nationaldutyrepaymentcenter.models.{AmendContent, UploadedFile}
 
 final case class AmendClaimRequest(Content: AmendContent, uploadedFiles: Seq[UploadedFile])
 
@@ -25,8 +25,5 @@ object AmendClaimRequest {
 
   implicit val formats: Format[AmendClaimRequest] =
     Json.format[AmendClaimRequest]
-
-  implicit val validate: Validator.Validate[AmendClaimRequest] =
-    Validator.always
 
 }
