+++ app/uk/gov/hmrc/nationaldutyrepaymentcenter/models/requests/CreateClaimRequest.scala
@@ -17,7 +17,7 @@
 package uk.gov.hmrc.nationaldutyrepaymentcenter.models.requests
 
 import play.api.libs.json.{Format, Json}
-import uk.gov.hmrc.nationaldutyrepaymentcenter.models.{Content, UploadedFile, Validator}
+import uk.gov.hmrc.nationaldutyrepaymentcenter.models.{Content, UploadedFile}
 
 final case class CreateClaimRequest(Content: Content, uploadedFiles: Seq[UploadedFile])
 
@@ -25,8 +25,5 @@ object CreateClaimRequest {
 
   implicit val formats: Format[CreateClaimRequest] =
     Json.format[CreateClaimRequest]
-
-  implicit val validate: Validator.Validate[CreateClaimRequest] =
-    Validator.always
 
 }
