+++ test/uk/gov/hmrc/nationaldutyrepaymentcenter/actors/FileTransferActorSpec.scala
@@ -16,6 +16,8 @@
 
 package uk.gov.hmrc.nationaldutyrepaymentcenter.actors
 
+import java.time.{LocalDateTime, ZonedDateTime}
+
 import akka.actor.{ActorSystem, Props}
 import akka.testkit.{ImplicitSender, TestActorRef, TestKit}
 import org.mockito.ArgumentMatchers.any
@@ -23,13 +25,12 @@ import org.mockito.Mockito.when
 import org.scalatest.wordspec.AnyWordSpecLike
 import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach}
 import org.scalatestplus.mockito.MockitoSugar.mock
-import uk.gov.hmrc.http.{HeaderCarrier, NotFoundException}
+import uk.gov.hmrc.http.{HeaderCarrier, UpstreamErrorResponse}
 import uk.gov.hmrc.nationaldutyrepaymentcenter.connectors.FileTransferConnector
 import uk.gov.hmrc.nationaldutyrepaymentcenter.models.{FileTransferResult, UploadedFile}
 import uk.gov.hmrc.nationaldutyrepaymentcenter.services.UUIDGenerator
 import uk.gov.hmrc.play.audit.http.connector.AuditConnector
 
-import java.time.{LocalDateTime, ZonedDateTime}
 import scala.concurrent.{ExecutionContext, Future}
 
 class FileTransferActorSpec
@@ -117,7 +118,7 @@ class FileTransferActorSpec
     "handle upstream errors" in {
 
       when(mockFileTransferConnector.transferFile(any())(any(), any()))
-        .thenReturn(Future.failed(new NotFoundException("File not found")))
+        .thenReturn(Future.failed(UpstreamErrorResponse.apply("Some error", 501)))
 
       val transferrer = system.actorOf(
         Props(
